Mk4/requirements.txt
requests
speedtest
opencv-python
psutil
pyautogui
pyjokes
pyttsx3
SpeechRecognition
wikipedia
winshell
selenium
translate
youtubetaggenerator
geocoder
pytube
numpy
pandas
cmake
face_recognition
screen-brightness-control
watchdog
uiautomation
pyaudio
chatterbot
moviepy
SpeechRecognition
pywave
pyfiglet
torch
bert-extractive-summarizer
pynput==1.7.1
pygame
pyshorteners